ax/langchain-py/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
chromadb>=0.6.0
pydantic>=2.0.0
arize-otel>=0.5.0
openinference-instrumentation-langchain>=0.1.0
